backend/requirements.txt
# SafeRoute Backend Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2

# Database
psycopg2-binary>=2.9.1
sqlalchemy>=1.4.23
asyncpg>=0.25.0

# Geospatial
geopy>=2.2.0
shapely>=1.7.1

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# API clients
requests>=2.25.1
aiohttp>=3.8.0
httpx>=0.18.0

# Utilities
python-dotenv>=0.19.0
email-validator>=1.1.3
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0

# Async support
asyncio>=3.4.3

# Testing
pytest>=6.2.4

# Development
black>=21.7b0
flake8>=3.9.2